refactor(atoms): merge single-tool atoms into grouped atoms + add contrib data packaging
Merge 9 single-tool atoms into 3 domain-grouped atoms:
- file_tools (read + write + edit): replaces tool_read, tool_write, tool_edit
- atom_management (install + unload + list): replaces tool_install_atom,
  tool_unload_atom, tool_list_atoms
- query_tools (traces + candidates + module_feedback): replaces
  tool_query_traces, tool_query_candidates, tool_query_module_feedback

LLM-facing tool names are unchanged. file_tools uses lazy-resolve for
Operations/ResourceWriter to avoid install-time ordering dependency on
the operations provider (file_tools sorts before operations_local
alphabetically; the old tool_read/tool_write sorted after it).

Update all scenario manifests, requires= declarations, test imports, and
comment references. Add [tool.uv.build-backend] data config to ship
contrib/scenarios/ in the wheel, and add sys.prefix lookup to the
scenario loader so wheel-installed scenarios are discoverable.
