+++ contrib/scenarios/rca/tuner/manifest.yaml
@@ -10,14 +10,10 @@ task_class: rca_baseline_tuner
 
 extensions:
   # File access for inspecting the current prompt + trace bodies.
-  - module: agentm.extensions.builtin.tool_read
+  - module: agentm.extensions.builtin.file_tools
 
   # Trace + candidate + feedback inspection.
-  - module: agentm.extensions.builtin.tool_query_traces
-  - module: agentm.extensions.builtin.tool_query_candidates
-    config:
-      default_scenario: rca:baseline
-  - module: agentm.extensions.builtin.tool_query_module_feedback
+  - module: agentm.extensions.builtin.query_tools
     config:
       default_scenario: rca:baseline
