+++ contrib/scenarios/sandbox/manifest.yaml
@@ -22,6 +22,5 @@ extensions:
     config:
       work_dir: /workspace
 
-  - module: agentm.extensions.builtin.tool_read
-  - module: agentm.extensions.builtin.tool_edit
+  - module: agentm.extensions.builtin.file_tools
   - module: agentm.extensions.builtin.tool_bash
