+++ pyproject.toml
@@ -67,6 +67,16 @@ sandbox = "agentm.scenarios.sandbox"
 requires = ["uv_build>=0.10.9,<0.11.0"]
 build-backend = "uv_build"
 
+# Ship contrib/scenarios/ inside the wheel so scenarios are discoverable
+# after ``pip install agentm``. The ``data`` category installs to
+# ``sys.prefix`` (the virtualenv root); the scenario loader's
+# ``_candidate_roots()`` includes ``sys.prefix`` as a search path, so
+# ``contrib/scenarios/<name>/manifest.yaml`` resolves from a wheel install
+# without ``AGENTM_PROJECT_ROOT`` or a source checkout.
+[tool.uv.build-backend]
+data = { data = "contrib" }
+source-include = ["contrib/scenarios/**"]
+
 [tool.uv.workspace]
 # Explicitly list nested projects under contrib/. Other contrib/scenarios/*
 # entries are pure manifests with no pyproject.toml and must NOT be picked
