+++ src/agentm/core/abi/extension.py
@@ -351,8 +351,8 @@ class ExtensionAPI(Protocol):
         """Replace the session's ``ResourceWriter``.
 
         Must be called at most once before freeze; a second call raises
-        ``KeyError``. Atoms that want to redirect ``tool_write`` /
-        ``tool_edit`` / ``tool_propose_change`` writes to a non-default
+        ``KeyError``. Atoms that want to redirect ``write`` /
+        ``edit`` / ``tool_propose_change`` writes to a non-default
         target (e.g. a sandbox-backed filesystem) call this at install
         time. If no atom calls it, ``get_resource_writer()`` returns the
         default :class:`~agentm.core.runtime.resource_writer.GitBackedResourceWriter`
